requirements.txt
django==1.5.1
python-yubikey==0.1.0
